Back-port continued: support up to 64 concurrent readers.

+++ src/tuple.h
@@ -45,7 +45,7 @@ public:
 
   fat_ptr clsn;     // version creation stamp
 #if defined(USE_PARALLEL_SSN) || defined(USE_PARALLEL_SSI)
-  typedef unsigned int rl_bitmap_t;  // _builtin_ctz needs it to be uint
+  typedef unsigned long rl_bitmap_t;  // use _builtin_ctzl for this
   rl_bitmap_t rl_bitmap;   // bitmap of in-flight readers
   fat_ptr sstamp;          // successor (overwriter) stamp (\pi in ssn), set to writer XID during
                            // normal write to indicate its existence; become writer cstamp at commit
